git-svn-id: http://springhead.info/spr2/Springhead2/trunk@3942 05cee5c3-a2e9-0310-9523-9dfc2f93dbe1

+++ src/Creature/CRReachingController.h
@@ -109,9 +109,9 @@ public:
 			CRReachingControllerIf*  reachCtl = creature->CreateController(reachDesc)->Cast();
 			controllers.push_back(reachCtl);
 			
-			for (size_t i=0; i<creature->NBodies(); ++i) {
+			for (int i=0; i<creature->NBodies(); ++i) {
 				CRBodyIf* body = creature->GetBody(i);
-				for (size_t j=0; j<body->NControlPoints(); ++j) {
+				for (int j=0; j<body->NControlPoints(); ++j) {
 					PHIKPosCtlIf* posCtl = body->GetControlPoint(j)->Cast();
 					if (posCtl && posCtl->GetSolid() == solid) {
 						reachCtl->SetIKCP(posCtl);
